requirements.txt
# ── Core API ──────────────────────────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.5.0
pyyaml>=6.0.0
httpx>=0.27.0
aiofiles>=24.0.0

# ── LLM Providers ─────────────────────────────────────────────────────────────
openai>=1.50.0
anthropic>=0.40.0
ollama>=0.4.0

# ── Voice – STT ───────────────────────────────────────────────────────────────
faster-whisper>=1.0.0
sounddevice>=0.4.6
soundfile>=0.12.0
numpy>=1.24.0

# ── Voice – Wake Word ─────────────────────────────────────────────────────────
openwakeword>=0.6.0

# ── Voice – TTS ───────────────────────────────────────────────────────────────
edge-tts>=6.1.0
pygame>=2.5.0

# ── Memory & Profile ──────────────────────────────────────────────────────────
chromadb>=0.5.0
aiosqlite>=0.20.0

# ── Phase 1 – Integrations (existing) ─────────────────────────────────────────
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
discord.py>=2.3.0
youtube-transcript-api>=0.6.0
requests>=2.31.0

# ── Phase 2 – New Core Integrations ───────────────────────────────────────────
# Google Calendar & Drive use same google-api-python-client (already included)
python-telegram-bot>=21.0.0
spotipy>=2.23.0
slack-sdk>=3.27.0
# whatsapp-web.py — install manually if needed: pip install whatsapp-web.py

# ── Phase 3 – Computer & Web Control ──────────────────────────────────────────
playwright>=1.45.0
pyautogui>=0.9.54
mss>=9.0.0
Pillow>=10.0.0
pytesseract>=0.3.10

# ── Phase 4 – Intelligence ────────────────────────────────────────────────────
pypdf>=4.0.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0

# ── Phase 5 – Platform & Ecosystem ───────────────────────────────────────────
PyGithub>=2.3.0
notion-client>=2.2.0
# plaid-python>=14.0.0  # uncomment to enable finance
# resemblyzer>=0.1.4    # uncomment to enable speaker recognition

# ── Scheduling ────────────────────────────────────────────────────────────────
apscheduler>=3.10.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0

# ── Utils ─────────────────────────────────────────────────────────────────────
websockets>=12.0
rich>=13.0.0
click>=8.1.0
